Add a simple omp target parallel for test case
The output files are generated by the latest dev branch of REX compiler. Currently, the driver and kernel functions are both stored in the same .cu file.
However, it failed to compile and execute the code correctly.

By split the driver and kernel to two files, the code can run without any problems.